Removed check if value has changed before updating devices
I integrated this check initially, since it seemed not necessary to update e.g. a valve position if it has nog changed. But I noticed that the graphs start looking empty if a sensor is not updated for a long time (which is the case with e.g. valve sensors during summertime ;-). After removing this check, the graphs look normal again.